+++ src/main.js
@@ -3,6 +3,9 @@ import App from "./App.vue";
 import router from "@/router";
 import VueRouter from "vue-router";
 
+// 3rd Party Plugins
+import '@/libs/toastification'
+
 // Import Stylesheets
 import "@/assets/css/styles.css";