+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added card and modal ui components
 - Added text input and data table ui components
+- Added toastification plugin for toast messages
 
 ### Changed